requirements.txt
streamlit>=1.33
pandas>=2.0
beautifulsoup4>=4.12
lxml>=4.9
requests>=2.31
playwright>=1.45
sqlalchemy>=2.0
streamlit-vis-timeline==0.3.0
openpyxl>=3.1.0
plotly>=5.22
python-dateutil>=2.9
streamlit-plotly-events
numpy==1.26.4
streamlit-plotly-events==0.0.6
streamlit-js-eval==0.1.5
